seyed-yaar-final/backend/requirements.txt
numpy>=1.24
shapely>=2.0
pillow>=10.0
rasterio>=1.3
requests>=2.31
python-dateutil>=2.8

python-dotenv>=1.0
copernicusmarine>=2.0
h5py>=3.10
netCDF4>=1.6
